project/requirements.txt
celery==4.4.1
Django==3.0.4
redis==3.4.1
psycopg2-binary
djangorestframework
django-filter
requests
pytesseract
pyyaml
uritemplate
ipython
django-cors-headers
python-docx
Pillow
python-magic
PyPDF2
fitz
PyMuPDF
drf-extra-fields
wand==0.5.7
